+++ src/components/Calendar/Calendar.module.css
@@ -3,6 +3,7 @@
   flex-direction: column;
   align-items: center;
   gap: 24px;
+  position: relative;
 }
 
 .calendarHead {
@@ -93,4 +94,24 @@
   .calendarItem {
     width: calc((100% - 301px) / 8);
   }
+}
+
+
+
+.loaderBg{
+  position: absolute;
+  top:0;
+  left:0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(255, 255, 255, 0.4);
+  z-index: 2;
+}
+.monthDataLoader{
+  position: absolute;
+  top:50%;
+  left:50%;
+  transition: transformY(50%);
+  transition: transformX(50%);
+  background-color: aqua;
 }
