+++ src/components/Loader/Loader.jsx
@@ -1,0 +1,8 @@
+import css from './Loader.module.css';
+const Loader = ({addClass = ""}) => {
+  return (
+    <div className={[css.loader, addClass].join(' ')}></div>
+  )
+}
+
+export default Loader
